requirements.txt
pygame
websocket